requirements.txt
mysql~=0.0.3
mysql-connector-python~=9.1.0
pandas~=2.2.3
PyMySQL~=1.1.1
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1